x86/cpufeatures: Add SGX1 and SGX2 sub-features
Add a feature word to hold SGX features enumerated via CPUID.0x12.0x0,
along with flags for SGX1 and SGX2.  As part of virtualizing SGX, KVM
needs to expose the SGX CPUID leafs to its guest.  SGX1 and SGX2 need to
be in a dedicated feature word so that they can be queried via KVM's
reverse CPUID lookup to properly emulate the expected guest behavior.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>